Refactor Zero Code Change Test Directory (#215)
