+++ tests/zero_code_change/horovod_tests/constants.py
@@ -1,0 +1,5 @@
+HOROVOD_KERAS_TEST_SCRIPT_PATH = "horovod_keras_mnist.py"
+HOROVOD_KERAS_TEST_SCRIPT_ARGS = ["--num_epochs", "2"]
+
+HOROVOD_ESTIMATOR_TEST_SCRIPT_PATH = "horovod_estimator_mnist.py"
+HOROVOD_ESTIMATOR_TEST_SCRIPT_ARGS = ["--num_steps", "1000"]
