+++ tests/zero_code_change/horovod_tests/tensorflow/test_estimator.py
@@ -2,30 +2,36 @@
 import os
 
 # Third Party
+from tests.zero_code_change.horovod_tests.constants import (
+    HOROVOD_ESTIMATOR_TEST_SCRIPT_ARGS,
+    HOROVOD_ESTIMATOR_TEST_SCRIPT_PATH,
+)
+from tests.zero_code_change.horovod_tests.utils import launch_horovod_job
 from tests.zero_code_change.utils import build_json
 
 # First Party
 from smdebug.trials import create_trial
 
 # Local
-from .utils import get_available_gpus, launch_horovod_job
+from .utils import get_available_gpus
+
 
 """
 Tested on current DLAMI p3.8xlarge when run from the main directory
 """
 
-HOROVOD_MNIST_SCRIPT_NAME = "horovod_estimator_mnist.py"
-HOROVOD_MNIST_ARGS = ["--num_steps", "1000"]
-
 
 def basic_test(out_dir, mode):
     path = build_json(out_dir, include_workers="one", include_collections=["weights", "gradients"])
     num_workers = len(get_available_gpus())
-    mode_args = list(HOROVOD_MNIST_ARGS) + ["--model_dir", os.path.join(out_dir, "checkpoint")]
+    mode_args = list(HOROVOD_ESTIMATOR_TEST_SCRIPT_ARGS) + [
+        "--model_dir",
+        os.path.join(out_dir, "checkpoint"),
+    ]
     if mode == "cpu":
         mode_args += ["--use_only_cpu", "true"]
     launch_horovod_job(
-        script_file_path=f"examples/tensorflow/sagemaker_official_container/{HOROVOD_MNIST_SCRIPT_NAME}",
+        script_file_path=f"examples/tensorflow/sagemaker_official_container/{HOROVOD_ESTIMATOR_TEST_SCRIPT_PATH}",
         script_args=mode_args,
         num_workers=num_workers,
         config_file_path=path,
@@ -50,11 +56,14 @@ def test_gpu(out_dir):
 def mode_allworkers(out_dir, mode):
     path = build_json(out_dir, include_workers="all", include_collections=["weights", "gradients"])
     num_workers = len(get_available_gpus())
-    mode_args = list(HOROVOD_MNIST_ARGS) + ["--model_dir", os.path.join(out_dir, "checkpoint")]
+    mode_args = list(HOROVOD_ESTIMATOR_TEST_SCRIPT_ARGS) + [
+        "--model_dir",
+        os.path.join(out_dir, "checkpoint"),
+    ]
     if mode == "cpu":
         mode_args += ["--use_only_cpu", "true"]
     launch_horovod_job(
-        script_file_path=f"examples/tensorflow/sagemaker_official_container/{HOROVOD_MNIST_SCRIPT_NAME}",
+        script_file_path=f"examples/tensorflow/sagemaker_official_container/{HOROVOD_ESTIMATOR_TEST_SCRIPT_PATH}",
         script_args=mode_args,
         num_workers=num_workers,
         config_file_path=path,
@@ -79,11 +88,14 @@ def mode_allworkers_saveall(out_dir, mode):
         out_dir, include_workers="all", save_all=True, include_collections=["weights", "gradients"]
     )
     num_workers = len(get_available_gpus())
-    mode_args = list(HOROVOD_MNIST_ARGS) + ["--model_dir", os.path.join(out_dir, "checkpoint")]
+    mode_args = list(HOROVOD_ESTIMATOR_TEST_SCRIPT_ARGS) + [
+        "--model_dir",
+        os.path.join(out_dir, "checkpoint"),
+    ]
     if mode == "cpu":
         mode_args += ["--use_only_cpu", "true"]
     launch_horovod_job(
-        script_file_path=f"examples/tensorflow/sagemaker_official_container/{HOROVOD_MNIST_SCRIPT_NAME}",
+        script_file_path=f"examples/tensorflow/sagemaker_official_container/{HOROVOD_ESTIMATOR_TEST_SCRIPT_PATH}",
         script_args=mode_args,
         num_workers=num_workers,
         config_file_path=path,
